floats

+++ components/AddCategory.tsx
@@ -61,7 +61,7 @@ const AddCategory = ({
         service_id: serviceId,
         type_Of_Repair: formData.typeOfRepair,
         raw: parseInt(formData.raw),
-        tax: parseInt(formData.tax),
+        tax: parseFloat(formData.tax),
         shipping: parseInt(formData.shipping),
         labour: parseInt(formData.labour),
         timeFrame: formData.timeFrame,
